Fix quantity being parsed as float rather than int

+++ server/controllers/inventoryController.js
@@ -26,7 +26,7 @@ export const getProduct = async (req, res, next) => {
 export const createProduct = async (req, res, next) => {
     const newProduct = {
         name: req.body.name,
-        quantity: parseFloat(req.body.quantity),
+        quantity: parseInt(req.body.quantity),
         unit: req.body.unit,
         price: parseFloat(req.body.price)
     }
@@ -59,7 +59,7 @@ export const updateProduct = async (req, res, next) => {
 
     await db.query(
         "UPDATE inventory SET name = ?, quantity = ?, unit = ?, price = ? WHERE product_id = ?",
-        [req.body.name, parseFloat(req.body.quantity), req.body.unit, parseFloat(req.body.price), id]
+        [req.body.name, parseInt(req.body.quantity), req.body.unit, parseFloat(req.body.price), id]
     );
     const [updatedInventory] = await db.query("SELECT * FROM inventory");
     res.status(200).json(updatedInventory);
